requirements.txt
Flask==2.3.3
Werkzeug<3.0
flask-restx==1.3.2
flask-jwt-extended==4.5.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
